third_party/WebKit/LayoutTests/platform/win/fast/mediacapturefromelement/HTMLMediaElementCapture-EME-content-expected.txt

CONSOLE ERROR: Uncaught (in promise) NotSupportedError: None of the requested configurations were supported.
This is a testharness.js-based test.
PASS check <video> captureStream() fails on an encrypted/protected media 
Harness: the test ran to completion.
